Code analysis warnings fixed (or suppressed it depends)

+++ Web.Upnp.Control/Services/Commands/QueueClearCommandHandler.cs
@@ -8,7 +8,7 @@ using Web.Upnp.Control.Services.Abstractions;
 
 namespace Web.Upnp.Control.Services.Commands
 {
-    public class QueueClearCommandHandler : IAsyncCommandHandler<QClearCommand>
+    public sealed class QueueClearCommandHandler : IAsyncCommandHandler<QClearCommand>
     {
         private readonly IUpnpServiceFactory factory;
 
@@ -19,6 +19,7 @@ namespace Web.Upnp.Control.Services.Commands
 
         public async Task ExecuteAsync(QClearCommand command, CancellationToken cancellationToken)
         {
+            if(command is null) throw new ArgumentNullException(nameof(command));
             var service = await factory.GetServiceAsync<QueueService>(command.DeviceId, cancellationToken).ConfigureAwait(false);
             var result = await service.RemoveAllAsync(0, command.QueueId, 0, cancellationToken).ConfigureAwait(false);
             await service.RemoveAllAsync(0, command.QueueId, uint.Parse(result["NewUpdateID"], CultureInfo.InvariantCulture), cancellationToken).ConfigureAwait(false);
